requirements.txt
Django>=4.2,<5.0
tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0

